Refactor worker and connector handlers into focused modules
